third_party/blink/web_tests/external/wpt/webrtc-stats/getStats-remote-candidate-address-expected.txt

This is a testharness.js-based test.
[FAIL] Do not expose in stats remote addresses that are not known to be already exposed to JS
  assert_equals: address should be null expected (object) null but got (string) ""
Harness: the test ran to completion.
